+++ public/javascripts/main.js
@@ -2,6 +2,7 @@ import { $, createNewElement, deleteClassFromElement } from './utils.js';
 import { createGenreNav, changeClickedNavTab } from './genre-nav.js';
 import { headerNavData } from './data.js';
 import { renderWebtoonPage } from './webtoon.js';
+import { renderWebNovelPage } from './web-novel.js';
 
 const HEADER_NAV_LIST = [
   { pageIndex: 0, pageName: '홈', renderFunction: renderHomePage },
@@ -49,16 +50,16 @@ function renderHomePage(pageIndex) {
   );
 }
 
-function renderWebNovelPage(pageIndex) {
-  const contentTag = $('.content');
-  contentTag.innerText = '';
-  changeClickedNavTab(
-    headerNavData,
-    pageIndex,
-    HEADER_NAV_LIST_CLASS_NAME,
-    HEADER_NAV_CLICKED_ITEM_CLASS_NAME
-  );
-}
+// function renderWebNovelPage(pageIndex) {
+//   const contentTag = $('.content');
+//   contentTag.innerText = '';
+//   changeClickedNavTab(
+//     headerNavData,
+//     pageIndex,
+//     HEADER_NAV_LIST_CLASS_NAME,
+//     HEADER_NAV_CLICKED_ITEM_CLASS_NAME
+//   );
+// }
 
 function renderMoviePage(pageIndex) {
   const contentTag = $('.content');
