+++ views/index.html
@@ -12,6 +12,7 @@
     <script type="module" src="javascripts/data.js"></script>
     <script type="module" src="javascripts/genre-nav.js"></script>
     <script type="module" src="javascripts/webtoon.js"></script>
+    <script type="module" src="javascripts/web-novel.js"></script>
     <script type="module" src="javascripts/main.js"></script>
     <div class="header">
       <div class="header__logo">kakaopage</div>
